feat(metrics): add cache_read_input_tokens tracking for CSA-lite (#1432)
Task 6 of CSA-lite Phase 1: surface Anthropic prompt-cache hit metrics on
session result output and JSON payload.

Schema
- StreamingMetadata (csa-core + csa-acp): add cache_read_input_tokens +
  cache_hit_ratio() helper.
- TokenUsage (csa-session::state): add cache_read_input_tokens with
  serde(default, skip_serializing_if = Option::is_none) for backwards
  compatibility, plus cache_hit_ratio() helper.

Population
- parse_token_usage: recognise "cache_read_input_tokens" before the
  shorter "input_tokens" probe and add a lookback guard so the longer
  key cannot shadow input_tokens.
- update_cumulative_tokens: only accumulate cache_read when present, so a
  None response from a non-Claude tool does not zero prior totals.
- convert_acp_metadata: bridge the new field across the ACP -> core
  StreamingMetadata boundary alongside existing input/output_tokens.

Display
- csa session result: print "Cache read: N tokens (P% hit rate)" line
  and expose total_token_usage.cache_hit_ratio in JSON payload.
- New load_total_token_usage helper reads state.toml directly so
  cross-project sessions render correctly.

Tests
- StreamingMetadata cache_hit_ratio: happy path 200K/150K -> 0.75, plus
  None on missing cache_read, missing input_tokens, and zero input.
- parse_token_usage_with_cache_read_input_tokens and the cache_read-only
  variant verify the lookback guard.

Module split
- session_cmds_result.rs and csa-acp/src/client.rs grew past the 800-line
  monolith cap; extracted handle_session_measure into
  session_cmds_result_measure.rs and the no-verify shell heuristic into
  csa-acp/src/client/no_verify_detect.rs. No behavior change.
